// src/types/enums.ts
// Mock, could include various frequencies
export enum PrescriptionFrequencies {
	'28day' = '28day',
	'56day' = '56day',
	'custom' = 'custom',
	'asNeeded' = 'asNeeded',
}

// Mock, could include various strenghts
export enum MedicineStrengths {
	TwoHundredFifty = '250',
	FiveHundred = '500',
	EightFifty = '850',
}

// Mock, could include various forms
export enum MedicineForms {
	powder = 'powder',
	infusionBags = 'infusionBags',
	oralSuspension = 'oralSuspension',
	tablets = 'tablets',
	suspension = 'suspension',
}